Ensure that Task repetitions are created only in the future
The idea of task repetitions is to repeat a task each n seconds.
This is fine for use-cases where task processing is always running.

When task processing is not always running, this can create a problem.
For example, imagine a daily task and the processing stopping for one week.
When the processing resumes, the same task will be run 7 times in short succession.

This commit ensures that task repetitions are only scheduled in the future.
For above example, this means that the 7 tasks will not be created.
Instead, only one new task will be created, preventing the same task
from running unnecessarily several times.

This commit also adds a unit test to ensure this behavior will not break
again in the future.