Move models_completed content to models.py
Django 2.1 (at least, not sure when it changed) doesn't detect the model
this way, and wants to create a migration to drop CompletedTask.

This leads on to failing tests when your app runs on PostgreSQL: The
test suite will try to TRUNCATE all tables after each test.
But since there's a table that Django doesn't know about (namely
CompletedTask), it will not be in the SQL statement to do this.
PostgreSQL then refuses to execute TRUNCATE.

If there's no specific reason to have this model in it's own file, we
should move it to avoid complications like this.