yes/no instead of green/red icons
In the django admin when a tasks failed the 'has_error' field shows
a green icon (and when the task succeeded a red icon).

So if for instance you are looking at a screen with all succesful tasks
you are greeted by a sea of red icons (somehow suggestive of problems).

This commit introduces a new BACKGROUND_TASK_BOOLEAN_AS_TEXT optional
setting. When set to True the icons are replaced by the words 'Yes' and
'No'.